refactor(mcp): extract reindex state manager into separate module
- Extract createReindexStateManager into reindex-state-manager.ts
- Remove code duplication from test file
- Export ReindexState from state manager module
- Update imports across server.ts, types.ts, and tests

Fixes Veille complexity violation:
- Test file no longer contains complex implementation (was 62 complexity)
- Better separation of concerns
- Reusable state manager module
