+++ packages/cli/src/mcp/server.reindex-state.test.ts
@@ -1,4 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { createReindexStateManager } from './reindex-state-manager.js';
 
 // Mock console.warn to verify warning logs
 const originalWarn = console.warn;
@@ -12,80 +13,6 @@ beforeEach(() => {
 afterEach(() => {
   console.warn = originalWarn;
 });
-
-/**
- * Reindex state manager implementation (extracted for testing)
- */
-function createReindexStateManager() {
-  interface ReindexState {
-    inProgress: boolean;
-    pendingFiles: string[];
-    lastReindexTimestamp: number | null;
-    lastReindexDurationMs: number | null;
-  }
-  
-  let state: ReindexState = {
-    inProgress: false,
-    pendingFiles: [],
-    lastReindexTimestamp: null,
-    lastReindexDurationMs: null,
-  };
-  
-  let activeOperations = 0;
-
-  return {
-    getState: () => ({ ...state }),
-    
-    startReindex: (files: string[]) => {
-      if (!files || files.length === 0) {
-        return;
-      }
-      
-      activeOperations += 1;
-      state.inProgress = true;
-      
-      // Merge new files into pending list (avoid duplicates)
-      const existing = new Set(state.pendingFiles);
-      for (const file of files) {
-        if (!existing.has(file)) {
-          state.pendingFiles.push(file);
-        }
-      }
-    },
-    
-    completeReindex: (durationMs: number) => {
-      if (activeOperations === 0) {
-        console.warn('[Lien] completeReindex called without matching startReindex');
-        return;
-      }
-      
-      activeOperations -= 1;
-      
-      // Only mark complete when all operations finish
-      if (activeOperations === 0) {
-        state.inProgress = false;
-        state.pendingFiles = [];
-        state.lastReindexTimestamp = Date.now();
-        state.lastReindexDurationMs = durationMs;
-      }
-    },
-    
-    failReindex: () => {
-      if (activeOperations === 0) {
-        console.warn('[Lien] failReindex called without matching startReindex');
-        return;
-      }
-      
-      activeOperations -= 1;
-      
-      // Only clear when all operations complete/fail
-      if (activeOperations === 0) {
-        state.inProgress = false;
-        state.pendingFiles = [];
-      }
-    },
-  };
-}
 
 describe('Reindex State Manager', () => {
   describe('basic state transitions', () => {
